Store file hashes when installing
This is part of a series of changes I am planning to make to handle updating
modpacks. It is important to me that plenty of considerations are made:
1. If a player has overriden a config, avoid replacing it.
2. If a player has installed a new mod, avoid removing it.
3. If a mod has been updated or removed, remove the old mod.

Equally we do need to update mods, however ftbinstall is going to take the
approach of many package managers - let the player handle updating their
configurations. We will store any files we cannot install under .ftbinstall
and write a note to the console, so that players can investigate and make
the changes they need.