technic: Install packs to the Minecraft launcher
Currently this supports packs that have a bin/version.json, but not
(quite) those with bin/modpack.jar files - I'm still working on it.

I'm also mulling over whether to use Technic's LaunchWrapper, or
whether this project should have its own legacy launcher (as older FTB
packs will also require this), or whether its fine to just use Mojang's
(which at this time I'm hopeful will work).

I also need to consider how to approach newer Forge versions and
Technic packs - we already have a Forge installer system, so being able
to use that (and install Forge the proper way) would be preferable. It
may be the case that redesigns to said Forge module may need to occur
to handle this use-case. For the time being, this is low-priority.