technic: fix jar merging
This commit is just fixes for the previous commit - while I believe
this should work, it doesn't. The Minecraft launcher doesn't like the
size of the version jar and downloads the client again :(

This commit resolves 2 problems, and the jar merger should now produce
the desired outcome:
1. Close the zip writer before the file, so that everything is written
   to file. This produces an actually valid zip file.
2. Filter out META-INF of the original jar, rather than everything but
   META-INF.

I'm going to continue to explore using this jar merging method for the
Minecraft launcher, as it should work (its how I believe Forge
operates). I also want to explore whether it would be possible to
simply put bin/modpack.jar at the start of the classpath (though from
my earlier experiments with using a library, I'm doubtful).

I'm also continuing to consider how to approach Minecraft Forge for
versions >= 1.6 - with the intention of being able to use the existing
Forge module to handle this, with some modifications of course.